% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution des trajets par jour de la semaine}
\usage{
calcul_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame.}
}
\value{
Un data.frame groupé par jour avec le total des trajets.
}
\description{
Calculer la distribution des trajets par jour de la semaine
}
