% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{Jeu de données des trajets vélos à Nantes (Toussaint 2025)}
\format{
Un data.frame avec les colonnes de comptage habituelles.
}
\source{
\url{https://data.nantesmetropole.fr/}
}
\usage{
df_velo
}
\description{
Données de comptage issues de l'Open Data de Nantes Métropole.
}
\keyword{datasets}
