% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les anomalies de trajet}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les données de comptage.}
}
\value{
Un data.frame sans les lignes contenant des anomalies.
}
\description{
Filtrer les anomalies de trajet
}
