% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Afficher le graphique de la distribution hebdomadaire}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame.}
}
\value{
Un objet ggplot représentant la distribution.
}
\description{
Afficher le graphique de la distribution hebdomadaire
}
