% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver la boucle avec le plus grand nombre de trajet}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data.frame.}
}
\value{
Un data.frame avec la ligne ayant le maximum de passages.
}
\description{
Trouver la boucle avec le plus grand nombre de trajet
}
