% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets par boucle}
\usage{
filtrer_trajet(trajet, boucle)
}
\arguments{
\item{trajet}{Un data.frame de données (ex: df_velo).}

\item{boucle}{Un vecteur de numéros de boucles.}
}
\value{
Le data.frame filtré.
}
\description{
Filtrer les trajets par boucle
}
